groq_streamlit/requirements.txt
streamlit==1.35.0
groq==0.9.0
streamlit-ace==0.1.1
python-dotenv==1.0.1
